Don't warn about installing foreign files twice
Add ability to suppress the warning about installing "foreign" files,
and use this when checking for excluded files. This way we don't warn if
the file just ends up excluded, and don't warn twice (once when
computing the path used for exclusion checking, and again when computing
the final path with prefix(es) stripped).